+++ pkg/cmd/init.go
@@ -1,9 +1,7 @@
 package cmd
 
 import (
-	"errors"
 	"fmt"
-	"net/url"
 	"os"
 	"path/filepath"
 
@@ -63,6 +61,8 @@ func runInitCommand(cmd *cobra.Command, args []string) error {
 		_, err = runPrompt(prompt)
 
 		if err != nil {
+			// TODO: make all the errors lowercase to make Go static check happy, then have something higher up that capitalizes the first letter
+			// of errors from commands
 			return fmt.Errorf("Not overwriting %s", filePath)
 		}
 
@@ -74,49 +74,6 @@ func runInitCommand(cmd *cobra.Command, args []string) error {
 
 	cmd.Printf("Wrote configuration to %s\n", filePath)
 	return nil
-}
-
-type decoratePromptFunc func(*promptui.Prompt) *promptui.Prompt
-
-func newApplianceURLPrompt() *promptui.Prompt {
-	return &promptui.Prompt{
-		Label: "Enter the URL of your Conjur service",
-		Validate: func(input string) error {
-			if len(input) == 0 {
-				return errors.New("URL is required")
-			}
-
-			_, err := url.ParseRequestURI(input)
-			return err
-		},
-	}
-}
-
-func newAccountPrompt() *promptui.Prompt {
-	return &promptui.Prompt{
-		Label: "Enter your organization account name",
-		Validate: func(input string) error {
-			if len(input) == 0 {
-				return errors.New("Account is required")
-			}
-			return nil
-		},
-	}
-}
-
-func newFileExistsPrompt(filePath string) *promptui.Prompt {
-	return &promptui.Prompt{
-		Label:     fmt.Sprintf("File %s exists. Overwrite", filePath),
-		IsConfirm: true,
-	}
-}
-
-func runPrompt(prompt *promptui.Prompt) (userInput string, err error) {
-	userInput, err = prompt.Run()
-	if err != nil {
-		return "", err
-	}
-	return userInput, nil
 }
 
 // NewInitCommand initializes and configures the 'conjur init' command.
